Update requirements.txt and add readme
requirements.txt has a bogus package `discord` listed, however, PyCharm does not
understand that the module `discord` is actually brought in by `discord.py`
package and not `discord`. Thus `discord` package is added to the requirements
solely to quiet PyCharm down about not listing the "right" package in reqs.txt.